Move Types into new folder
